Phase 4: move Nextra emitter to @jolli.ai/site-core
Eighteen renderer files migrated, all pure (string templates + IR →
TemplateFile transforms; no fs / subprocess / network):

  cli/src/site/renderer/nextra/ApiCss.{ts,test.ts}
  cli/src/site/renderer/nextra/ApiNavMethods.ts
  cli/src/site/renderer/nextra/Components.{ts,test.ts}
  cli/src/site/renderer/nextra/EndpointDataEmitter.{ts,test.ts}
  cli/src/site/renderer/nextra/EndpointPageEmitter.{ts,test.ts}
  cli/src/site/renderer/nextra/OverviewPageEmitter.{ts,test.ts}
  cli/src/site/renderer/nextra/Paths.{ts,test.ts}
  cli/src/site/renderer/nextra/SidebarMetaEmitter.{ts,test.ts}
  cli/src/site/renderer/nextra/Types.ts
  cli/src/site/renderer/nextra/index.{ts,test.ts}

  → site-core/src/renderer/nextra/

`OpenApiSpecInput` definition relocates from
`cli/src/site/renderer/SiteRenderer.ts` to
`site-core/src/openapi/Types.ts` so the moved emitter files can reach
it from within site-core (the previous `../SiteRenderer.js` import
broke after the move). CLI's `SiteRenderer.ts` re-exports
`OpenApiSpecInput` from `@jolli.ai/site-core` for backward-compat with
CLI consumers that still import it from this module — the `SiteRenderer`
interface itself stays in CLI because it depends on the I/O-bound
`ServerResult` (NpmRunner) and `OutputFilter` types.

CLI consumers updated to import from `@jolli.ai/site-core`:

  cli/src/commands/StartCommand.ts        — buildApiSidebarOverrides
  cli/src/site/renderer/NextraRenderer.ts — generateApiCss,
                                             EMPTY_API_NAV_METHODS,
                                             emitNextraOpenApiFiles,
                                             generateApiComponents,
                                             TemplateFile

`site-core/src/index.ts` adds the renderer surface to the public API
(`export * from "./renderer/nextra/index.js"` + the two ApiCss helpers
that nextra/index.ts doesn't re-export).

After the move, `cli/src/site/renderer/` contains only:
  - SiteRenderer.ts        (interface, depends on CLI-only I/O types)
  - NextraRenderer.{ts,test.ts}  (the I/O class implementing SiteRenderer)
  - index.{ts,test.ts}     (renderer factory — picks NextraRenderer)

`cli/src/site/renderer/nextra/` directory removed (empty after move).

Verification:
  - site-core: 23 test files, 475 tests, coverage 99.23% / 96.55% /
    100% / 99.55% (all ≥ 97/96/97/97)
  - cli: all tests passing; coverage 98.68% / 97.58% / 98.02% / 98.85%
  - vscode: unaffected
  - `npm run all` clean end-to-end
  - VS Code esbuild bundle still compiles (the renderer/nextra/ files
    were not imported from `vscode/src/`)
