+++ site-core/src/renderer/nextra/EndpointDataEmitter.ts
