+++ site-core/src/renderer/nextra/SidebarMetaEmitter.ts
